feat: mover dashboard SDR para /dashboard (Visão Geral)
- dashboard/page.tsx substituído pelo conteúdo do SDR IA dashboard (KPIs, funil, sentimento, WhatsApp, sessões)
- sdr-ia/dashboard/page.tsx: redirect server-side para /dashboard
- sdr-ia/layout.tsx: remove tab Dashboard (link agora é na sidebar)
- Sidebar.tsx: SDR IA aponta para /sdr-ia/parametros em vez de /sdr-ia/dashboard
